requirements.txt
boto3==1.33.7
botocore==1.33.7
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
jmespath==1.0.1
python-dateutil==2.8.2
python-decouple==3.8
requests==2.31.0
s3transfer==0.8.2
six==1.16.0
urllib3==2.0.7
